+++ data.tf
@@ -17,19 +17,13 @@ data "aws_iam_policy_document" "container_instance" {
   }
 }
 
-data "template_cloudinit_config" "config" {
-  gzip          = false
-  base64_encode = true
-
-  # Cloud-config configuration for installing ssm agent.
-  part {
-    content_type = "text/x-shellscript"
-    content      = templatefile("${path.module}/scripts/userdata.sh", {})
-  }
-
-  # Additional script
-  part {
-    content_type = "text/x-shellscript"
-    content      = var.extra_script
-  }
+locals {
+  userdata_script = base64encode(join("\n", [
+    templatefile("${path.module}/scripts/userdata.sh", {
+      ssm_temp_dir              = var.ssm_agent_temp_directory
+      ssm_agent_version_centos6 = var.ssm_agent_version_centos6
+      ssm_urls                  = var.ssm_agent_s3_urls
+    }),
+    var.extra_script
+  ]))
 }