+++ main.tf
@@ -17,7 +17,7 @@ resource "aws_ecs_cluster" "this" {
 
         content {
           kms_key_id = var.create_kms_key ? aws_kms_key.main[0].key_id : var.kms_key_id
-          logging    = lookup(execute_command_configuration.value, "logging", "DEFAULT")
+          logging    = lookup(execute_command_configuration.value, "logging", var.default_execute_command_logging)
 
           dynamic "log_configuration" {
             for_each = try([execute_command_configuration.value.log_configuration], [])
@@ -59,4 +59,5 @@ resource "aws_kms_key" "main" {
   enable_key_rotation     = var.enable_key_rotation
   policy                  = local.kms_policy
   deletion_window_in_days = var.deletion_window_in_days
+  tags                    = var.tags
 }