+++ outputs.tf
@@ -23,3 +23,13 @@ output "key_id" {
   description = " The globally unique identifier for the key."
   value       = aws_kms_key.main.*.key_id
 }
+
+output "autoscaling_group_arn" {
+  description = "ARN of the Auto Scaling Group"
+  value       = try(aws_autoscaling_group.container_instance[0].arn, null)
+}
+
+output "autoscaling_group_name" {
+  description = "Name of the Auto Scaling Group"
+  value       = try(aws_autoscaling_group.container_instance[0].name, null)
+}