+++ versions.tf
@@ -4,12 +4,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 5.0.0"
-    }
-
-    template = {
-      source  = "hashicorp/template"
-      version = ">= 2.0.0"
+      version = ">= 5.0.0, < 6.0.0"
     }
   }
 }